updates, need to do optimization or template specialization for more optimized synchronization...basically it works now but can be much better

+++ main.cpp
@@ -6,10 +6,10 @@
 #include "ringbuffer.tcc"
 
 struct Data{
-   Data( int64_t send ) : send_count(  send )
+   Data( size_t send ) : send_count(  send )
    {}
 
-   int64_t                 send_count;
+   size_t                 send_count;
 } data( 1e7 ) ;
